requirements.txt
apache-airflow==2.1.2
apache-airflow-providers-postgres==1.0.2
apache-airflow-providers-http==1.1.0
apache-airflow-providers-sftp==1.2.0
apache-airflow-providers-ssh==1.3.0

#apache-airflow-providers-google==3.0.0
#apache-airflow-providers-imap==1.0.1
#psycopg2-binary==2.9.1
